+++ agape-core/agape/files/tests.py
@@ -1,43 +1,54 @@
-
-# testing classes
 from django.test import TestCase
 from rest_framework.test import APIRequestFactory
 from rest_framework.test import APIClient
 
-# agape classes
-from agape.signals import on
-
-# utility classes
 import json
+from django.core.files.uploadedfile import SimpleUploadedFile
+from django.conf import settings
 from django.utils import timezone
 from datetime import timedelta
-import datetime
 
-# package classes
-from .models import Group
-from .serializers import GroupSerializer
+import os.path
+import shutil
+
+# Create your tests here.
+from .models import File
+from .serializers import FileSerializer
+
+from agape.people.models import Person
+from agape.people.serializers import PersonSerializer
+
+from agape.signals import on
 
 
-MODEL = Group
-SERIALIZER_CLASS = GroupSerializer
+MODEL = File
+SERIALIZER_CLASS = FileSerializer
 CREATE_DATA =  {   
-        'name':'Blue Man Group',
-        'description': 'Create the ultimate dance party.'
+        'name': 'blue-man-file.pdf',
+        'type': 'pdf',
+        'path': 'aso0w98we098ds09.pdf',
+
+        'label': 'Blue Man File',
+        'description': 'Guide to creating the perfect person.',
+
     }
 EXPECT_DATA = {   
     'id':1,
-    'name':'Blue Man Group',
-    'description': 'Create the ultimate dance party.',
     'progenitor': None,
-    'status':0
-}
+    'name': 'blue-man-file.pdf',
+    'type': 'pdf',
+    'path': 'aso0w98we098ds09.pdf',
+
+    'label': 'Blue Man File',
+    'description': 'Guide to creating the perfect person.',
+    }
 
 UPDATE_DATA = { 
     'description': 'I am that I am.'
 }
 UPDATE_EXPECT_DATA= {   
     'id':1,
-    'name':'Blue Man Group',
+    'label':'Blue Man File',
     'description': 'I am that I am.'
 }
 
@@ -138,23 +149,33 @@ class SerializerTestCase(TestCase):
             instance_value = getattr(instance, key)
             self.assertEqual(instance_value, expected_value, "Comparing {} attribute".format(key) )
 
-
-
 class APITestCase(TestCase):
 
     def setUp(self):
         self.model              = MODEL
-        self.serializer_class   = CREATE_DATA
-        self.expect_data        = EXPECT_DATA
-        self.update_data        = UPDATE_DATA
+        self.serializer_class   = SERIALIZER_CLASS
+
+        self.create_data        = CREATE_DATA.copy()
+        del self.create_data['path']
+        self.create_data['file'] =  SimpleUploadedFile("file.mp4", b"file_content", content_type="video/mp4")
+
+        self.expect_data        = EXPECT_DATA.copy()
+        del self.expect_data['path']
+        self.update_data        = UPDATE_DATA.copy()
+
         self.update_expect_data = UPDATE_EXPECT_DATA
 
         self.client =  APIClient()
         self.entity = 'group'
-        self.api_end_point = '/api/v1/groups/'
+        self.api_end_point = '/api/v1/files/'
+
+    def test_no_file_supplied(self):
+        response = self.client.post(self.api_end_point, self.create_data)
+        self.assertEqual(response.status_code, 400, "Created new instance")
+
 
     def test_create(self):
-        response = self.client.post(self.api_end_point, self.create_data)
+        response = self.client.post(self.api_end_point, self.create_data, format='multipart')
         self.assertEqual(response.status_code, 201, "Created new instance")
 
         # compare response to expected data
@@ -166,8 +187,11 @@ class APITestCase(TestCase):
         instance = self.model.objects.get(id=response.data.get('id'))
         self.assertTrue(instance)
 
+        # verify that the uploaded file exists on disk
+        self.assertTrue(os.path.isfile(instance.full_path()));
+
     def test_retrieve(self):
-        response = self.client.post(self.api_end_point, self.create_data)
+        response = self.client.post(self.api_end_point, self.create_data, format='multipart')
         self.assertEqual(response.status_code, 201, "Created new instance")
 
         instance_id = response.data.get('id')
@@ -182,11 +206,27 @@ class APITestCase(TestCase):
             instance_value = response.data.get(key)
             self.assertEqual(instance_value, expected_value, "Comparing {} attribute".format(key) )
 
+    def test_download(self):
+        response = self.client.post(self.api_end_point, self.create_data, format='multipart')
+        self.assertEqual(response.status_code, 201, "Created new instance")
+
+        # TODO: Test file download
+
+        instance_id = response.data.get('id')
+        uri = "{}{}/".format(self.api_end_point,instance_id)
+        print(uri)
+
+        response = self.client.get(uri)
+
+        self.assertEqual(response.status_code, 200, "Created new instance")
+
+
     def test_update(self):
-        response = self.client.post(self.api_end_point, self.create_data)
+        response = self.client.post(self.api_end_point, self.create_data, format='multipart')
         self.assertEqual(response.status_code, 201, "Created new instance")
 
         instance_id = response.data.get('id')
+        original_instance = File.objects.get(id=instance_id)
         uri = "{}{}/".format(self.api_end_point,instance_id)
 
         response = self.client.patch(uri, json.dumps(self.update_data), content_type='application/json')
@@ -199,6 +239,7 @@ class APITestCase(TestCase):
 
         # perform get operation and validate
         instance_id = response.data.get('id')
+        instance = File.objects.get(id=instance_id)
         uri = "{}{}/".format(self.api_end_point,instance_id)
 
         response = None
@@ -209,11 +250,15 @@ class APITestCase(TestCase):
             instance_value = response.data.get(key)
             self.assertEqual(instance_value, expected_value, "Comparing {} attribute".format(key) )
 
+        # self.assertTrue(original_instance.path != instance.path )
+
     def test_delete(self):
-        response = self.client.post(self.api_end_point, self.create_data)
+        response = self.client.post(self.api_end_point, self.create_data, format='multipart')
         self.assertEqual(response.status_code, 201, "Created new instance")
 
+
         instance_id = response.data.get('id')
+        instance = File.objects.get(id=instance_id)
         uri = "{}{}/".format(self.api_end_point,instance_id)
 
         response = self.client.delete( uri )    
@@ -222,10 +267,9 @@ class APITestCase(TestCase):
         query = self.model.objects.filter(id=instance_id)
         self.assertEqual(len(query),0, "Deleted")
 
+        # verify that the uploaded file exists on disk
+        self.assertTrue(not os.path.isfile(instance.full_path()));
 
-# class OrganizationGroupsTestCase(TestCase):
-
-#     def setup():
-#         pass
-
-#     def 
+    def tearDown(self):
+        if (os.path.isdir(settings.MEDIA_ROOT)):
+            shutil.rmtree(settings.MEDIA_ROOT)