+++ agape-core/tests/urls.py
@@ -8,4 +8,5 @@ urlpatterns = [
     url(r'^api/v1/', include('agape.groups.urls') ),
     url(r'^api/v1/', include('agape.members.urls') ),
     url(r'^api/v1/', include('agape.events.urls') ),
+    url(r'^api/v1/', include('agape.files.urls') )
 ]