+++ agape-core/CHANGES.md
@@ -1,3 +1,10 @@
+
+v0.3.0 April 25
+- Added files application with upload support
+
+v0.2.9 April 24
+- All email links support https://
+
 v0.2.8 April 24
 - Activation email links support https://